Refactor to allow CLI options and fix the server.
The README states the port the server is ran on "can be changed by
passing the port as an option to the server". However, this does not
work. Upon inspection, I found that the SERVER and PORT are hard coded
in constants.js, and these values then used everywhere, thus not
allowing you to set the port. Additionally, the SERVER should be
settable via the CLI.

In light of this, I refactored constants.js, and renamed it to
config.js, since it will no longer contain only constants. I used the
lightweight minimist package to ease the parsing of the command line
arguments.

The server and port can now be set using the new -p|--port and
-s|--server options. I also exposed the raw command ARGUMENTS and
OPTIONS for use in the built-in clients.

I also updated the README to reflect these changes.
