--- src/constants.js
@@ -1,7 +1,0 @@
-module.exports = {
-    PORT: 8082,
-    SERVER: 'localhost',
-
-    MSG_ID: 'msg_id',
-    MSG_EVENT: 'msg_event'
-}
